implemented displaying initial introduction for a book
This includes a bunch of infrastructure for future development:
 - Reader: the interface shown to the player
 - Executor: the engine that manages the book's universe while
   it is being played
 - Book: the content being executed.

The Executor sends content updates to the Reader. The content
includes potential actions the Reader can send back to the
Executor to be interpreted by the book content.

Unlike most redux applications, this one has two Store objects:
one for the Reader state, and one for the Executor state.
Eventually the Reader and Executor will operate in separate
contexts, so they will need their own Store objects. The Reader
will operate in the main browser thread and update the DOM,
while the executor will operate in a protected worker thread.
The executor could even be operating on a server instead.